refactor: make `ArtworkListImageBorder` reusable
