+++ src/app/Components/ArtworkLists/components/ArtworkListImageBorder.tsx
